Major cleanup. Now connect (agent) uses tokio::UdpSocket, instead of socket2.
