+++ mirrord-layer/src/outgoing/tcp.rs
@@ -2,35 +2,35 @@ use std::{
     collections::HashMap,
     future::Future,
     net::{IpAddr, Ipv4Addr, Ipv6Addr, SocketAddr},
-    ops::{Deref, DerefMut},
 };
 
 use futures::SinkExt;
 use mirrord_protocol::{
-    outgoing::udp::{DaemonUdpOutgoing, LayerUdpOutgoing},
+    outgoing::{tcp::*, DaemonConnect, DaemonRead, LayerClose, LayerConnect, LayerWrite},
     ClientCodec, ClientMessage, ConnectionId,
 };
 use tokio::{
-    net::UdpSocket,
+    io::{AsyncReadExt, AsyncWriteExt},
+    net::{TcpListener, TcpStream},
     select,
     sync::mpsc::{channel, Receiver, Sender},
     task,
 };
 use tokio_stream::{wrappers::ReceiverStream, StreamExt};
-use tracing::{debug, error, info, trace, warn};
+use tracing::{error, info, trace, warn};
 
 use super::*;
 use crate::{common::ResponseDeque, detour::DetourGuard, error::LayerError};
 
 /// Hook messages handled by `TcpOutgoingHandler`.
 #[derive(Debug)]
-pub(crate) enum UdpOutgoing {
+pub(crate) enum TcpOutgoing {
     Connect(Connect),
 }
 
 /// Responsible for handling hook and daemon messages for the outgoing traffic feature.
 #[derive(Debug)]
-pub(crate) struct UdpOutgoingHandler {
+pub(crate) struct TcpOutgoingHandler {
     /// Holds the channels used to send daemon messages to the interceptor socket, for the case
     /// where (agent) received data from the remote host, and sent it to (layer), to finally be
     /// passed all the way back to the user.
@@ -44,32 +44,11 @@ pub(crate) struct UdpOutgoingHandler {
     /// main `layer` loop.
     ///
     /// This is sent from `interceptor_task`.
-    layer_tx: Sender<LayerUdpOutgoing>,
-    layer_rx: Receiver<LayerUdpOutgoing>,
+    layer_tx: Sender<LayerTcpOutgoing>,
+    layer_rx: Receiver<LayerTcpOutgoing>,
 }
 
-/// Wrapper type around `tokio::Sender`, used to send messages from the `agent` to our interceptor
-/// socket, where they'll be written back to the user's socket.
-///
-/// (agent) -> (layer) -> (user)
-#[derive(Debug)]
-pub(crate) struct ConnectionMirror(tokio::sync::mpsc::Sender<Vec<u8>>);
-
-impl Deref for ConnectionMirror {
-    type Target = tokio::sync::mpsc::Sender<Vec<u8>>;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-impl DerefMut for ConnectionMirror {
-    fn deref_mut(&mut self) -> &mut Self::Target {
-        &mut self.0
-    }
-}
-
-impl Default for UdpOutgoingHandler {
+impl Default for TcpOutgoingHandler {
     fn default() -> Self {
         let (layer_tx, layer_rx) = channel(1000);
 
@@ -82,15 +61,15 @@ impl Default for UdpOutgoingHandler {
     }
 }
 
-impl UdpOutgoingHandler {
+impl TcpOutgoingHandler {
     async fn interceptor_task(
-        layer_tx: Sender<LayerUdpOutgoing>,
+        layer_tx: Sender<LayerTcpOutgoing>,
         connection_id: ConnectionId,
-        mirror_socket: UdpSocket,
+        mut mirror_stream: TcpStream,
         remote_rx: Receiver<Vec<u8>>,
     ) {
         let mut remote_stream = ReceiverStream::new(remote_rx);
-        let mut recv_from_buffer = vec![0; 1500];
+        let mut buffer = vec![0; 1024];
 
         // Sends a message to close the remote stream in `agent`, when it's
         // being closed in `layer`.
@@ -99,22 +78,20 @@ impl UdpOutgoingHandler {
         // `read`ing.
         let close_remote_stream = |layer_tx: Sender<_>| async move {
             let close = LayerClose { connection_id };
-            let outgoing_close = LayerUdpOutgoing::Close(close);
+            let outgoing_close = LayerTcpOutgoing::Close(close);
 
             if let Err(fail) = layer_tx.send(outgoing_close).await {
                 error!("Failed sending close message with {:#?}!", fail);
             }
         };
 
-        // TODO(alex) [low] 2022-09-07: Connect this socket to the user socket.
-        let mut user_address: Option<SocketAddr> = None;
-
         loop {
             select! {
                 biased; // To allow local socket to be read before being closed
 
-                read = mirror_socket.recv_from(&mut recv_from_buffer) => {
-                    debug!("read from recv_from");
+                // Reads data that the user is sending from their socket to mirrord's interceptor
+                // socket.
+                read = mirror_stream.read(&mut buffer) => {
                     match read {
                         Err(fail) if fail.kind() == std::io::ErrorKind::WouldBlock => {
                             continue;
@@ -125,19 +102,17 @@ impl UdpOutgoingHandler {
 
                             break;
                         }
-                        Ok((read_amount, _)) if read_amount == 0 => {
+                        Ok(read_amount) if read_amount == 0 => {
                             info!("interceptor_task -> Stream {:#?} has no more data, closing!", connection_id);
                             close_remote_stream(layer_tx.clone()).await;
 
                             break;
                         },
-                        Ok((read_amount, from)) => {
-                            debug!("from {:#?}", from);
-                            user_address = Some(from);
+                        Ok(read_amount) => {
                             // Sends the message that the user wrote to our interceptor socket to
                             // be handled on the `agent`, where it'll be forwarded to the remote.
-                            let write = LayerWrite { connection_id, bytes: recv_from_buffer[..read_amount].to_vec() };
-                            let outgoing_write = LayerUdpOutgoing::Write(write);
+                            let write = LayerWrite { connection_id, bytes: buffer[..read_amount].to_vec() };
+                            let outgoing_write = LayerTcpOutgoing::Write(write);
 
                             if let Err(fail) = layer_tx.send(outgoing_write).await {
                                 error!("Failed sending write message with {:#?}!", fail);
@@ -153,13 +128,7 @@ impl UdpOutgoingHandler {
                             // Writes the data sent by `agent` (that came from the actual remote
                             // stream) to our interceptor socket. When the user tries to read the
                             // remote data, this'll be what they receive.
-                            if let Err(fail) = mirror_socket
-                                .send_to(
-                                    &bytes,
-                                    user_address.expect("User socket should be set by now!"),
-                                )
-                                .await
-                            {
+                            if let Err(fail) = mirror_stream.write_all(&bytes).await {
                                 error!("Failed writing to mirror_stream with {:#?}!", fail);
                                 break;
                             }
@@ -181,14 +150,14 @@ impl UdpOutgoingHandler {
 
     /// Handles the following hook messages:
     ///
-    /// - `UdpOutgoing::Connect`: inserts the new connection request into a connection queue, and
-    ///   sends it to (agent) as a `UdpOutgoingRequest::Connect` with the remote host's address.
+    /// - `TcpOutgoing::Connect`: inserts the new connection request into a connection queue, and
+    ///   sends it to (agent) as a `TcpOutgoingRequest::Connect` with the remote host's address.
     ///
-    /// - `UdpOutgoing::Write`: sends a `UdpOutgoingRequest::Write` message to (agent) with the data
+    /// - `TcpOutgoing::Write`: sends a `TcpOutgoingRequest::Write` message to (agent) with the data
     ///   that our interceptor socket intercepted.
     pub(crate) async fn handle_hook_message(
         &mut self,
-        message: UdpOutgoing,
+        message: TcpOutgoing,
         codec: &mut actix_codec::Framed<
             impl tokio::io::AsyncRead + tokio::io::AsyncWrite + Unpin + Send,
             ClientCodec,
@@ -197,17 +166,10 @@ impl UdpOutgoingHandler {
         trace!("handle_hook_message -> message {:?}", message);
 
         match message {
-            UdpOutgoing::Connect(Connect {
+            TcpOutgoing::Connect(Connect {
                 remote_address,
                 channel_tx,
             }) => {
-                // TODO(alex) [mid] 2022-09-06: We need to check if this `remote_address` is
-                // actually a local address! If it is, then the `agent` won't be able to reach it.
-                // Right now we're sidestepping this issue by just changing the address in `agent`
-                // to be a "local-agent" address.
-                //
-                // Have to be careful, as this is rather finnicky behavior when the user wants to
-                // make a request to their own local address.
                 trace!("Connect -> remote_address {:#?}", remote_address);
 
                 // TODO: We could be losing track of the proper order to respond to these (aviram
@@ -215,7 +177,7 @@ impl UdpOutgoingHandler {
                 self.connect_queue.push_back(channel_tx);
 
                 Ok(codec
-                    .send(ClientMessage::UdpOutgoing(LayerUdpOutgoing::Connect(
+                    .send(ClientMessage::TcpOutgoing(LayerTcpOutgoing::Connect(
                         LayerConnect { remote_address },
                     )))
                     .await?)
@@ -225,25 +187,25 @@ impl UdpOutgoingHandler {
 
     /// Handles the following daemon messages:
     ///
-    /// - `UdpOutgoingResponse::Connect`: grabs the reply from the connection request that was sent
-    ///   to (agent), then creates a new `UdpListener` (the interceptor socket) that the user socket
+    /// - `TcpOutgoingResponse::Connect`: grabs the reply from the connection request that was sent
+    ///   to (agent), then creates a new `TcpListener` (the interceptor socket) that the user socket
     ///   will connect to. When everything succeeds, it spawns a new task that handles the
     ///   communication between user and interceptor sockets.
     ///
-    /// - `UdpOutgoingResponse::Read`: (agent) received some data from the remote host and sent it
+    /// - `TcpOutgoingResponse::Read`: (agent) received some data from the remote host and sent it
     ///   back to (layer). The data will be sent to our interceptor socket, which in turn will send
     ///   it back to the user socket.
     ///
-    /// - `UdpOutgoingResponse::Write`: (agent) sent some data to the remote host, currently this
+    /// - `TcpOutgoingResponse::Write`: (agent) sent some data to the remote host, currently this
     ///   response is only significant to handle errors when this send failed.
     pub(crate) async fn handle_daemon_message(
         &mut self,
-        response: DaemonUdpOutgoing,
+        response: DaemonTcpOutgoing,
     ) -> Result<(), LayerError> {
         trace!("handle_daemon_message -> message {:?}", response);
 
         match response {
-            DaemonUdpOutgoing::Connect(connect) => {
+            DaemonTcpOutgoing::Connect(connect) => {
                 trace!("Connect -> connect {:#?}", connect);
 
                 let DaemonConnect {
@@ -251,30 +213,33 @@ impl UdpOutgoingHandler {
                     remote_address,
                 } = connect?;
 
-                let mirror_socket = {
+                let mirror_stream = {
                     let _ = DetourGuard::new();
 
-                    let mirror_socket = match remote_address {
+                    let mirror_listener = match remote_address {
                         SocketAddr::V4(_) => {
-                            UdpSocket::bind(SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 0))
+                            TcpListener::bind(SocketAddr::new(IpAddr::V4(Ipv4Addr::LOCALHOST), 0))
                                 .await?
                         }
                         SocketAddr::V6(_) => {
-                            UdpSocket::bind(SocketAddr::new(IpAddr::V6(Ipv6Addr::UNSPECIFIED), 0))
+                            TcpListener::bind(SocketAddr::new(IpAddr::V6(Ipv6Addr::UNSPECIFIED), 0))
                                 .await?
                         }
                     };
 
                     // Creates the listener that will wait for the user's socket connection.
-                    let mirror_address = MirrorAddress(mirror_socket.local_addr()?);
+                    let mirror_address = MirrorAddress(mirror_listener.local_addr()?);
 
                     self.connect_queue
                         .pop_front()
-                        .ok_or(LayerError::SendErrorUdpResponse)?
+                        .ok_or(LayerError::SendErrorTcpResponse)?
                         .send(Ok(mirror_address))
-                        .map_err(|_| LayerError::SendErrorUdpResponse)?;
+                        .map_err(|_| LayerError::SendErrorTcpResponse)?;
 
-                    mirror_socket
+                    // Accepts the user's socket connection, and finally becomes the interceptor
+                    // socket.
+                    let (mirror_stream, _) = mirror_listener.accept().await?;
+                    mirror_stream
                 };
 
                 let (remote_tx, remote_rx) = channel::<Vec<u8>>(1000);
@@ -284,16 +249,16 @@ impl UdpOutgoingHandler {
 
                 // user and interceptor sockets are connected to each other, so now we spawn a new
                 // task to pair their reads/writes.
-                task::spawn(UdpOutgoingHandler::interceptor_task(
+                task::spawn(TcpOutgoingHandler::interceptor_task(
                     self.layer_tx.clone(),
                     connection_id,
-                    mirror_socket,
+                    mirror_stream,
                     remote_rx,
                 ));
 
                 Ok(())
             }
-            DaemonUdpOutgoing::Read(read) => {
+            DaemonTcpOutgoing::Read(read) => {
                 // (agent) read something from remote, so we write it to the user.
                 trace!("Read -> read {:?}", read);
                 let DaemonRead {
@@ -308,7 +273,7 @@ impl UdpOutgoingHandler {
 
                 Ok(sender.send(bytes).await?)
             }
-            DaemonUdpOutgoing::Close(connection_id) => {
+            DaemonTcpOutgoing::Close(connection_id) => {
                 // (agent) failed to perform some operation.
                 trace!("Close -> connection_id {:?}", connection_id);
                 self.mirrors.remove(&connection_id);
@@ -320,7 +285,7 @@ impl UdpOutgoingHandler {
 
     /// Helper function to access the channel of messages that are to be passed directly as
     /// `ClientMessage` from `layer`.
-    pub(crate) fn recv(&mut self) -> impl Future<Output = Option<LayerUdpOutgoing>> + '_ {
+    pub(crate) fn recv(&mut self) -> impl Future<Output = Option<LayerTcpOutgoing>> + '_ {
         self.layer_rx.recv()
     }
 }
